moved HAL_Shim* to new directory HAL
